move schedule components to their own folder
